+++ src/main/java/org/zalando/nakadi/controller/EventPublishingController.java
@@ -131,9 +131,6 @@ public class EventPublishingController implements NakadiProblemHandling {
             final long msSpent = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startingNanos);
             final String applicationName = client.getClientId();
 
-            LOG.info("[SLO] [publishing-latency] time={} size={} count={} eventTypeName={} app={}", msSpent,
-                    totalSizeBytes, eventCount, eventTypeName, applicationName);
-
             nakadiKpiPublisher.publish(kpiBatchPublishedEventType, () -> new JSONObject()
                     .put("event_type", eventTypeName)
                     .put("app", applicationName)